+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added `ProtocolVersion.V8` corresponding to Protocol version 8
 - Added `BakerSuspended` and `BakerResumed` transaction result events
 - Added `GetConsensusDetailedStatus` endpoint for querying detailed consensus status information
+- Added `GetScheduledReleaseAccounts` endpoint for querying the list of accounts that have scheduled releases
 
 ## 8.0.0
 - Support for Protocol 7